CocoResort/CocoChat/requirements.txt
Flask
twilio
python-dotenv
openai
tiktoken
